Use LiveView form for internal performance editing
