+++ lib/jumubase_web/templates/internal/performance/edit.html.eex
@@ -6,7 +6,13 @@
 
 <hr>
 
-<%= render JumubaseWeb.PerformanceView, "_form.html",
+<%= live_render(@conn, JumubaseWeb.Internal.PerformanceLive.Edit, session: %{
+  "contest_id" => @contest.id,
+  "performance_id" => @performance.id,
+  "submit_title" => gettext("Save Changes")
+  }) %>
+
+<%# render JumubaseWeb.PerformanceView, "_form.html",
   conn: @conn,
   path: Routes.internal_contest_performance_path(@conn, :update, @contest, @performance),
   changeset: @changeset,
